testing framework set up, server file in src

+++ server/src/controllers/user.controller.ts
@@ -2,8 +2,10 @@ import UserSchema from "../schema/user.schema";
 import type { Request, Response } from "express";
 import bcrypt from "bcrypt";
 import { handleError } from "../utils";
-import { AuthRequest } from "../types/express";
+import { AuthRequest } from "../../types/express";
+import { isSelfOrStaff } from "../utils/auth";
 
+// GET /api/users/
 export const getAllUsers = async (req: Request, res: Response) => {
   try {
     const users = await UserSchema.find({}, { password: 0 });
@@ -13,6 +15,7 @@ export const getAllUsers = async (req: Request, res: Response) => {
   }
 };
 
+// GET /api/users/:id
 export const getUser = async (req: Request<{ id: string }>, res: Response) => {
   const id = req.params.id;
   try {
@@ -24,6 +27,7 @@ export const getUser = async (req: Request<{ id: string }>, res: Response) => {
   }
 };
 
+// POST /api/users
 export const createUser = async (
   req: Request<{}, {}, { email: string; username: string; password: string }>,
   res: Response,
@@ -48,8 +52,9 @@ export const createUser = async (
   }
 };
 
+// PUT /api/users/:id
 export const updateUser = async (
-  req: Request<
+  req: AuthRequest<
     { id: string },
     {},
     {
@@ -67,26 +72,39 @@ export const updateUser = async (
   try {
     const user = await UserSchema.findById(id);
     if (!user) return res.status(400).json({ message: "User doesn't exist" });
-    const updatedUser = await UserSchema.findByIdAndUpdate(id, {
-      email: email ?? user.email,
-      username: username ?? user.username,
-      password: password ? await bcrypt.hash(password, 10) : user.password,
-      reputation: reputation ?? user.reputation,
-      isStaff: isStaff ?? user.isStaff,
-    });
+    const isAllowed = await isSelfOrStaff(req, user.id);
+    if (!isAllowed) return res.status(401).json({ message: "Forbidden" });
+
+    const updatedUser = await UserSchema.findByIdAndUpdate(
+      id,
+      {
+        email: email ?? user.email,
+        username: username ?? user.username,
+        password: password ? await bcrypt.hash(password, 10) : user.password,
+        reputation: reputation ?? user.reputation,
+        isStaff: isStaff ?? user.isStaff,
+      },
+      { projection: { password: 0 }, returnOriginal: false },
+    );
+
+    res.status(201).json(updatedUser);
   } catch (err) {
     handleError(err, res);
   }
 };
 
+// DELETE /api/users/:id
 export const deleteUser = async (
-  req: Request<{ id: string }>,
+  req: AuthRequest<{ id: string }>,
   res: Response,
 ) => {
   const id = req.params.id;
   try {
-    const user = await UserSchema.findByIdAndDelete(id);
+    const user = await UserSchema.findById(id);
     if (!user) return res.status(404).json({ message: "User not found" });
+    const isAllowed = await isSelfOrStaff(req, user.id);
+    if (!isAllowed) return res.status(401).json({ message: "Forbidden" });
+    await UserSchema.findByIdAndDelete(id);
     res.status(204).json({ message: "User successfully deleted" });
   } catch (err) {
     handleError(err, res);
